+++ CHANGELOG.md
@@ -1,3 +1,24 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 26-Aug-2019
+* Release version 0.0.1.
 
-* TODO: Describe initial release.
+## [0.0.2] - 27-Aug-2019
+* Added badgeText in entry
+* Removed badgeText from list tile
+
+## [0.0.3] - 27-Aug-2019
+* fixed badge width
+
+## [0.0.4] - 27-Aug-2019
+* fixed badge status issue
+
+## [0.0.5] - 28-Aug-2019
+* fixed batch swipe issue on 0 child
+
+## [0.0.6] - 28-Aug-2019
+* fixed tile background
+
+## [0.0.7] - 28-Aug-2019
+* fixed background color
+
+## [0.0.8] - 28-Aug-2019
+* added configuration for batch selection color